# tradequest_spec_v2/Architecture/Detailed_API_Spec.yaml
openapi: 3.0.3
info:
  title: TradeQuest API — Detailed Spec (MVP)
  version: 1.0.0
servers:
  - url: https://api.atqventures.com
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tq_session
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error: {type: string, example: "validation_error"}
        code: {type: integer, example: 422}
        message: {type: string, example: "Missing required field: symbol"}
        request_id: {type: string, example: "req_01HGWJ..."}
...
    StrategySpec:
      type: object
      required: [name]
      properties:
        name:
          type: string
          enum: [sma_cross, rsi_revert, atr_trail]
        params:
          type: object
          additionalProperties: true
          example: {"fast": 10, "slow": 50}
        fees_bps: {type: number, default: 2.0}
        slip_bps: {type: number, default: 2.0}
    BacktestResult:
      type: object
      required: [metrics,equity_curve,trades,monte_carlo]
      properties:
        metrics:
          type: object
          properties:
            total_return_pct: {type: number}
            sharpe_simple: {type: number}
            max_dd_pct: {type: number}
            win_rate: {type: number}
            trades: {type: integer}
        equity_curve:
          type: array
          items:
            type: object
            properties:
              ts: {type: string, format: date-time}
              eq: {type: number}
        trades:
          type: array
          items:
            type: object
            properties:
              entry_ts: {type: string, format: date-time}
              exit_ts: {type: string, format: date-time}
              entry: {type: number}
              exit: {type: number}
              rtn: {type: number}
        monte_carlo:
          type: object
          properties:
            median: {type: number}
            p05: {type: number}
            p95: {type: number}
security:
  - cookieAuth: []
paths:
  /auth/magic-link:
    post:
      tags: [auth]
      summary: Request magic link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: {type: string, format: email}
      responses:
        "200": {description: Sent}
        "429":
          description: Rate-limited
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /auth/consume:
    post:
      tags: [auth]
      summary: Exchange token for session
      responses:
        "200": {description: OK}
        "400":
          description: Invalid/expired
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
...
  /backtest/run:
    post:
      tags: [backtest]
      summary: Backtest on uploaded OHLCV
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [symbol, ohlcv_csv_base64, strategy]
              properties:
                symbol: {type: string}
                ohlcv_csv_base64: {type: string}
                strategy: {$ref: "#/components/schemas/StrategySpec"}
      responses:
        "200":
          description: Metrics & curves
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BacktestResult"}
  /backtest/quick:
    post:
      tags: [backtest]
      summary: Backtest on live-fetched OHLCV
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [symbol, tf, strategy]
              properties:
                symbol: {type: string}
                tf: {type: string, enum: ["1m","5m","15m","1h","4h","1d"]}
                lookback_bars: {type: integer, default: 2000}
                strategy: {$ref: "#/components/schemas/StrategySpec"}
      responses:
        "200":
          description: Metrics & curves
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BacktestResult"}